Replaced cli.svg, sized up "whatisgulp", changed code example
The logo at the top of the page now uses `.editor-window`, so it's a regular `<pre><code>` block. This means it will scale the same as the code example, and it will be copy-paste-able. Also added breakpoints to the code example, so it is fully visible at all screen widths without scrolling. Both the logo and code example scale from font-size: 14pt to 13pt on small screens since the code example was just a bit too big at 14pt.